tegra-rcm: Download code to tegra in recovery mode
This program is used to download BCT and code to tegra from recovery
mode. Only tegras that are configurated for "nvproduction" or "ODM
open" mode are supported (no encryption).

usage:

tegra-rcm --bct=bctfile --bootloader=blfile
--loadaddr=<loadaddr> --entryaddr=<entryaddr>

bctfile is the name of the BCT (boot configuration table). It
contains memory timing and boot device selection.

blfile is the name of the booloader file. This is the code that

downloaded and executed on the tegra SoC.

loadaddr and entryaddr are the load and entry addresses of the
bootloader in the tegra address map.

Change-Id: I34754d3976f3971bdca1daefc5d1a2c69a56bbb5
Reviewed-on: http://git-master/r/94264
Reviewed-by: Allen Martin <amartin@nvidia.com>
Tested-by: Allen Martin <amartin@nvidia.com>