Add option --usb-timeout=<value>
RCM communication sometimes fails if the USB timeout is short. The default
timeout is 1000ms. Increasing the timeout value may avoid this issue. The
exact cause is not yet diagnosed. This patch implements a --usb-timeout
option to allow the value to be configured. A value of 0 means unlimited
timeout.

Example:
$ sudo tegrarcm --bct=jetson-tk1-bct.bct --bootloader=u-boot.bin \
  --loadaddr=0x83d88000 --pkc=rsa_priv.der --usb-timeout=5000

Signed-off-by: Jimmy Zhang <jimmzhang@nvidia.com>
(swarren, commit description fixes)
Signed-off-by: Stephen Warren <swarren@nvidia.com>